Move DetailActions under actions folder
