+++ client/src/actions/DetailActions.js
